Implement X11 HW-synced swapping using glXJoinSwapGroupNV, glXBindSwapBarrierNV, and the existing traits
This change is meant to be analogous to commit 634344aef504 ("From Marius
Heise, "here is a patch that implements Win32 HW-synced swapping using
wglJoinSwapGroupNV, wglBindSwapBarrierNV and the existing traits. It was
tested with multiple ATI FirePro S400 cards.") for the X11 backend. Tested
using multiple NVIDIA A5000 cards. The NVIDIA driver stack requires that
the swap group join and bind calls occur on the same thread context that is
doing the swap buffer operation.

Signed-off-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>